move proxies to use NSG for inbound network rules
Security Lists have an intrinsic limit of 5 per subnet. This change
implements using Network Security Lists for control of proxy network
security posture.

We define one Network Security Group (NSG) for both proxies and
network_security_group.tf becomes the source for the rules governing
the NSG. The NSG is attached to the proxy1 and proxy2 VNIC.

Terraform will cycle the VNIC when doing this attachment, so if you
are applying this change to a running infra, it will cause the proxies
to have new public IP