Add sketch_graph.py and example.json
This is a prototype of a script to take the JSON specification
defined by RTI-55 and output a png graph and confluence table. The
script can be invoked as:

`sketch_graph.py example.json`

This will generate a png file called example.png and a confluence
markup table called example.conflu. The png graph can be inline
attached to a confluence page. The markup can be insterted with
"insert markup" and the "Confluence wiki" selected in the format
dropdown in the pop-up.

This requires python3 and graphviz to be installed.

Fixes RTI-81