Fixes RTI-177, RTI-178, RTI-179
You can backup boot volumes in object storage.
When standing up infrastructure a bash script is placed in the external/oci directory that when run will backup all hosts in the engagement.
If the user sets `backup_compartment_id` in their variables.tfvars then the images will be moved to an archive compartment