Consolidated all backflip payload generation and connection management to a single user interface
The original backflip code was just a couple of scripts to infect Linux, then it was extended with more scripts to generate payloads for macOS.
Recently we had the need for generating payloads for Windows and this highlighted the need for simplification.
To simplify the user experience we refactored the backflip as a sort of framework that generates payloads based on templates for each OS based on a centralized C2 configuration.